Move transaction request tests beside txn request code
